build,style: upgrade clang tooling to v21 and apply auto-fixes
Pin clang-format and clang-tidy to version 21 across CMake and CI:
- CMakeLists.txt: format/format-check now invoke clang-format-21; the lint
  python runner invokes run-clang-tidy-21.
- ci.yml: install clang-format-21 and clang-tidy-21 from apt.llvm.org
  since Ubuntu's default repos lag behind v21.

Run clang-format-21 across the codebase and apply every auto-fixable
clang-tidy-21 suggestion that is mechanically safe. The dominant
families touched are modernize-use-scoped-lock, modernize-use-
designated-initializers, modernize-use-ranges, modernize-use-starts-
ends-with, modernize-use-integer-sign-comparison, readability-container-
size-empty, readability-container-contains, and readability-use-concise-
preprocessor-directives.

Repair the handful of sites where clang-tidy's auto-fix produced a
syntactically-broken edit:
- chat_ui_input_controller.cpp, composer_state.cpp, command_palette.cpp,
  markdown/block_parser.cpp: extra ')' from cmp_less rewrite that did
  not strip the original static_cast<int>(...) right paren.
- chat_ui_overlay_state.cpp: fix replaced size_t with the unqualified
  size_type identifier; restore std::size_t.
- mcp_server_session.hpp, oauth/flow.hpp, test_mcp_bearer_auth.cpp,
  test_mcp_streamable_http.cpp: nodiscard fix emitted "static
  [[nodiscard]] T f()", which is ill-formed; reorder to "[[nodiscard]]
  static T f()".

All 859 tests pass on Linux post-change. ~470 lint warnings dropped to
114; remaining warnings are non-mechanical and tracked separately.

+++ src/presentation/chat_ui_overlay_state.cpp
@@ -6,6 +6,7 @@
 #include "ui_spacing.hpp"
 
 #include <algorithm>
+#include <cstddef>
 #include <string>
 #include <utility>
 
@@ -64,19 +65,19 @@ std::string PrettyPrintJson(std::string_view json, size_t max_bytes) {
         out += c;
         out += '\n';
         ++depth;
-        out.append(depth * 2, ' ');
+        out.append(static_cast<std::size_t>(depth * 2), ' ');
         break;
       case '}':
       case ']':
         out += '\n';
         --depth;
-        out.append(depth * 2, ' ');
+        out.append(static_cast<std::size_t>(depth * 2), ' ');
         out += c;
         break;
       case ',':
         out += c;
         out += '\n';
-        out.append(depth * 2, ' ');
+        out.append(static_cast<std::size_t>(depth * 2), ' ');
         break;
       case ':':
         out += ": ";
